requirements.txt
Django>=5.1
djangorestframework
djangorestframework-simplejwt
drf-yasg
drf-spectacular